First pass with pylint, removing all unused imports, renamed setup module to initialize to remove ambiguity
