+++ jobs/E3SMDiags.py
@@ -7,7 +7,7 @@ from pprint import pformat
 from datetime import datetime
 from shutil import copyfile
 
-from lib.events import Event_list
+from lib.events import EventList
 from lib.slurm import Slurm
 from JobStatus import JobStatus, StatusMap
 from lib.util import (render,
