+++ lib/test_runmanager.py
@@ -6,7 +6,7 @@ import shutil
 from configobj import ConfigObj
 from runmanager import RunManager
 from filemanager import FileManager
-from events import Event_list
+from events import EventList
 
 
 class TestRunManager(unittest.TestCase):
@@ -37,7 +37,7 @@ class TestRunManager(unittest.TestCase):
             remote_path=self.remote_path,
             local_endpoint=self.local_endpoint)
         runmanager = RunManager(
-            event_list=Event_list(),
+            event_list=EventList(),
             output_path=self.output_path,
             caseID=self.config['global']['experiment'],
             scripts_path=self.run_scripts_path,
@@ -85,7 +85,7 @@ class TestRunManager(unittest.TestCase):
             remote_path=self.remote_path,
             local_endpoint=self.local_endpoint)
         runmanager = RunManager(
-            event_list=Event_list(),
+            event_list=EventList(),
             output_path=self.output_path,
             caseID=self.config['global']['experiment'],
             scripts_path=self.run_scripts_path,
